Big one, quote responder modal on internal side working, changed jquery to built in js solution and split up the home controller into 3  different ones: HomeController, InternalResponseController, VendorCatalogueController
