+++ Application/Application.csproj
@@ -20,6 +20,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Folder Include="Views\VendorTab\" />
     <Folder Include="wwwroot\lib\tingle\" />
   </ItemGroup>
