+++ Application/Views/Home/VendorMenu.cshtml
@@ -49,10 +49,10 @@
 {
     <script type="text/javascript">
         document.querySelector('.cs-vendor').addEventListener('click', function() {
-            document.location.href = "@Url.Action("VendorCatalogue", "Home")";
+            document.location.href = "@Url.Action("index", "VendorCatalogue")";
         });
         document.querySelector('.cs-accounts').addEventListener('click', function() {
-            document.location.href = "@Url.Action("VendorQuotes", "Home")";
+            document.location.href = "@Url.Action("VendorQuotes", "InternalResponse")";
         });
     </script>
 }
