+++ Application/Views/InternalResponse/QuoteResponderModal.cshtml
@@ -1,0 +1,49 @@
+﻿@using Services.EntityFramework.DbEntities
+@model Services.Models.DTOs.QuoteDTO
+<div>
+    <div class="bg-white shadow-sm">
+        <div style="background-image: url(@Model.VendorItem.ItemImageURL)" class="vendor-response-header"></div>
+        <div class="p-3">
+            <h2>@Model.VendorItem.ItemName</h2>
+            <div>@Model.VendorItem.ItemDescription</div>
+            <div><b>Quantity Requested:</b> @Model.QuantityRequested</div>
+        </div>
+    </div>
+    <h5 class="text-black-50 mt-3">Previous Responses</h5>
+    <div>
+        @foreach (var response in Model.Responses)
+        {
+        <div class="p-2">
+            <div class="d-inline-block w-100 text-end">
+                <div class="status-box" style="background-color:@response.Status.Colour">@response.Status.StatusName</div>
+            </div>
+            <div class="row mb-2 p-2 pb-1 shadow-sm bg-white">
+                <div class="col-3">
+                    <h6 class="text-black-50 mb-0">Date</h6>
+                    <div>@response.ResponseDate.ToShortDateString()</div>
+                </div>
+                <div class="col-6">
+                    <h6 class="text-black-50 mb-0">Message</h6>
+                    <div>@response.ReponseText</div>
+                </div>
+                <div class="col-3">
+                    <h6 class="text-black-50 mb-0">Price Offered</h6>
+                    <div>£@response.QuotePrice</div>
+                </div>
+            </div>
+        </div>
+        }
+    </div>
+    @if (Model.Responses.Any() && Model.Responses.OrderByDescending(o => o.QuoteResponseId).First().Status.AsEnum == QuoteStatusEnum.AWAITING_REVIEW)
+    {
+        <h5 class="text-black-50 mt-3">Action </h5>
+        <div class="align-content-center">
+            <div class="btn btn-success" data-approve-btn>Approve</div>
+            <div class="btn btn-danger" data-reject-btn>Reject</div>
+        </div>
+    }
+    else if(!Model.Responses.Any())
+    {
+        <h5 class="text-black-50 mt-3" style="font-size:18px;">Vendor has not responded yet.</h5>
+    }
+</div>
