+++ Application/Views/InternalResponse/VendorQuotes.cshtml
@@ -49,7 +49,7 @@
                             </div>
                         </div>
                         <div class="col-2 p-3 text-end">
-                            <div class="btn-primary" data-quote-btn="@quote.QuoteID">Open Quote</div>
+                            <div class="btn-primary" data-responder-btn="@quote.QuoteID">Open Quote</div>
                         </div>
                     </div>
                 }
@@ -84,7 +84,7 @@
                                     </div>
                                 </div>
                                 <div class="col-2 p-3 text-end">
-                                    <div class="btn-primary" data-quote-btn="@quote.QuoteID">Open Quote</div>
+                                    <div class="btn-primary" data-responder-btn="@quote.QuoteID">Open Quote</div>
                                 </div>
                             </div>
                         }
@@ -101,9 +101,15 @@
 
 @section Scripts
 {
+    <script src="~/js/quote-responder-modal.js" asp-append-version="true"></script>
     <script type="text/javascript">
-        window.onload = function() {
+        window.onload = function () {
             document.querySelector('#tabApprovedQuotes').style.display = "none";
+
+            QuoteModal({
+                quoteResponderModal: '@Url.Action("QuoteResponderModal", "InternalResponse")',
+                respondToQuote: '@Url.Action("Respond", "InternalResponse")'
+            });
         };
 
         document.querySelector('#tabApproved').addEventListener('click', (t) => {
